% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAanalysis_additional-fxns.R
\name{CG}
\alias{CG}
\title{CG content function}
\usage{
CG(sequence)
}
\arguments{
\item{sequence}{a single vector of individual character}

\item{wordsize}{the length of characters to include in nucleotide "word" search; for CG this is 2}

\item{alphabet}{all characters within the seqence to analyze}
}
\description{
This function allows you calculate the fraction of C and G content of a specified sequence.
}
\examples{
CG(FISHsequence, wordsize= 2, alphabet= c("a", "c", "t", "g"))
}
\keyword{CG}
