% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAanalysis-package.R
\docType{package}
\name{DNAanalysis-package}
\alias{DNAanalysis}
\alias{DNAanalysis-package}
\title{DNAanalysis: An introduction to DNA analysis}
\description{
This package is an introduction to DNA manipulation and analysis.
    Specifically, this package covers the following:
        Reading in DNA sequences
        Simple Summaries (length, nucleotide content, CG and GC calculations, sliding plot window)
        Distance matrix calculation
        Phylogenetic construction
        Mosaic plot construction
}
\keyword{internal}
